Develop ruch

+++ src/main/java/com/efitops/basesetup/dto/DespatchInstructionDTO.java
@@ -24,8 +24,8 @@ public class DespatchInstructionDTO {
 	private Long location;
 	private String paymentTerms;
 	private String modeOfTransport;
-	private int netWeight;
-	private int grossWeight;
+	private double netWeight;
+	private double grossWeight;
 	private String deliveryInstructions;
 	private String Consignee;
 	private Long orgId;
